added support for expired windowed elements

+++ include/spatialDistance.h
@@ -1,5 +1,7 @@
 #pragma once
 #include "stdafx.h"
+#include <stdint.h>
+
 
 
 class spatialDistance
@@ -11,7 +13,7 @@ class spatialDistance
         double haversine(double lat1, double lon1, double lat2, double lon2);
         double equirectangular(double lat1, double lon1, double lat2, double lon2);
         double sphericalLawOfCosines(double lat1, double lon1, double lat2, double lon2);
-        int getNumDistanceComputations(void);
+        uint64_t getNumDistanceComputations(void);
 
 
 
@@ -19,7 +21,7 @@ class spatialDistance
 
     private:
         const double mEARTH_RADIUS = 6371008.7714; // Earth radius in meters
-        static int mNumDistanceComputations;
+        static uint64_t mNumDistanceComputations;
 
         double degreesToRadians(double degValue);
